requirements.txt
cycler==0.10.0
dj-database-url==0.5.0
Django==2.2.7
django-heroku==0.3.1
django-storages==1.8
et-xmlfile==1.0.1
jdcal==1.4.1
kiwisolver==1.1.0
matplotlib==3.1.2
numpy==1.17.4
openpyxl==3.0.2
pandas==0.25.3
Pillow==6.2.1
psycopg2==2.8.4
pybase64==1.0.1
pyparsing==2.4.5
pytesseract==0.3.0
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
whitenoise==4.1.4
